requirements.txt
clarifai
instabot
selenium
Appium-Python-Client
